[Prover Service] Refactor circuit input signals.
